# perf: add `Any` overload for `SyntaxTokenList`