Add `Any` overload for `SyntaxTokenList`, saves 0.5 MB on the complex code bencmark. 


### Benchmarks
#### Before
| Method                        | Mean     | Error   | StdDev   | Gen0      | Gen1      | Gen2      | Allocated |
|------------------------------ |---------:|--------:|---------:|----------:|----------:|----------:|----------:|
| Default_CodeFormatter_Complex | 241.7 ms | 7.22 ms | 21.17 ms | 7000.0000 | 4000.0000 | 1000.0000 |  61.61 MB |

#### After
| Method                        | Mean     | Error   | StdDev  | Gen0      | Gen1      | Gen2      | Allocated |
|------------------------------ |---------:|--------:|--------:|----------:|----------:|----------:|----------:|
| Default_CodeFormatter_Complex | 234.2 ms | 4.65 ms | 6.82 ms | 7000.0000 | 4000.0000 | 1000.0000 |  61.09 MB |